Alternative carbon allocation approach
Implemented an option carbon allocation routine (DailyPRTAllometricCarbonSimpler
in parteh/PRTAllometricCarbonMod.F90). In this routine, I applied much simpler rules
for maintenance allocation: it simply checks how much each pool is in deficit, and
allocates carbon (storage + carbon_balance) to the pools according to the debt. If
there is any carbon left, then the plant finishes filling the storage pool and then
allocates to growth. This option is set by fates_parteh_mode=0, and will require
changing the valid options in the host land models (see example of changes for
CTSM in here).

This was extracted from the now defunct PR #800. Kept as a separate pull request if
we ever want to bring it to the mainline.